create list and video component, render inside playlist component, wire up list component to playlist-reducer
